modubot/requirements.txt
aiosqlite
discord.py